"feat(exec_MVP_TESTE):  refactored to exec_com(). Now in norme. Should work propperly"
