+++ headers/defs.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   defs.h                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: viferrei <viferrei@student.42sp.org.br>    +#+  +:+       +#+        */
+/*   By: coder <coder@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/27 23:09:13 by coder             #+#    #+#             */
-/*   Updated: 2022/10/08 02:23:56 by coder            ###   ########.fr       */
+/*   Updated: 2022/10/09 21:22:04 by coder            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
